ril: Update ril.h, kill ril.mk
The updated ril.h file will fix compilation problems, but to fix
the dependency error that occurs on WiFi-only devices (which should
not exist), msm8916-common cannot use "BOARD_PROVIDES_LIBRIL".

All device trees will be required to call that flag (not common device
trees) if this change is merged. This will come in handy if any LTE
devices wish to remain supported and co-exist with WiFi-only devices.

Change-Id: I89f2568de312837605cba1c55aa93509bea81955