msm8916-common: Set ro.bt.bdaddr_path prop
In Pie, Google removed random MAC generation from the bluetooth
impl, and this is causing crash when enabling bluetooth on many
devices. Some devices, like ours, stores the MAC address in a
file in the efs partition.

Point ro.bt.bdaddr_path to this file, which fixes the bluetooth crash.

Change-Id: I2901b9af3af99ca71963778d55fa7ec82fe96b84