msm8916-common: rootdir: Don't use early init script
This stopped working on Q, making a service for it is a half-working solution, and the script doesn't do anything we can't do with makefiles. In fact, the latter is better, and lots of devices don't use early init scripts anymore. Power off alarm doesn't seem to work.

The necessary changes which were being done through this script will now be done from device-specific makefiles.

Signed-off-by: Gegham Zakaryan <zakaryan.2004@outlook.com>
Change-Id: Ie43e889645284b6a9103742c416ec6a52fcd3a99